% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feeds_urls.R
\name{get_data}
\alias{get_data}
\title{Retrieve Bikeshare Data from GBFS Feed}
\usage{
get_data(url)
}
\arguments{
\item{url}{URL of the GBFS feed to retrieve}
}
\value{
A list containing the data and last updated time stamp
}
\description{
Retrieve Bikeshare Data from GBFS Feed
}
\details{
This package provides functions to retrieve and process bike-sharing system data using the General Bikeshare Feed Specification (GBFS).
It allows users to easily fetch and analyze bikeshare information from various data feeds.
}
\examples{
\dontrun{
feeds_urls() \%>\%
  filter(name == "station_information") \%>\%
  pull(url) \%>\%
  get_data()
}
}
